Separate image data from sample images
* Only store image path and position info in sample images.
* Have a separate ImageData class to handle image data and metadata.
* This avoids automatically using a lot of memory from image data in
  the sample instance.
